Address Notes from PR 410
* Break Bus Performance Manager logic into multiple files
  - one chooses files
  - one converts gtfs rt into vehicle events
  - one has gtfs schedule utilities
* Fixes in Biz Logic for conversion from gtfs_rt into vehicle events
  - inclusive filtering on schedule utlities
  - cast all fields explicitly
  - keep all datetimes as utc
  - remove extra `unique` call
  - don't sort output, leave that to next part of the pipeline
  - handle potential edge cases where 'current_status' is different from
    the two common forms.
  - don't do any shifting of data, leave the for later.
